added lint

+++ contracts/MerkleTree.sol
@@ -2,6 +2,7 @@ pragma solidity ^0.5.0;
 pragma experimental ABIEncoderV2;
 import {MerkleTreeLib as MTLib} from "./libs/MerkleTreeLib.sol";
 
+
 /*
  * Merkle Tree Utilities for Rollup
  */
@@ -12,13 +13,16 @@ contract MerkleTree {
     struct MerkleTree {
         // The root
         bytes32 root;
-        uint height;
-        mapping (bytes32 => bytes32) nodes;
+        uint256 height;
+        mapping(bytes32 => bytes32) nodes;
     }
-    
+
     constructor(address _mtLibAddress) public {
         MerkleLib = MTLib(_mtLibAddress);
-        setMerkleRootAndHeight(MerkleLib.getZeroRoot(), MerkleLib.getMaxTreeDepth());
+        setMerkleRootAndHeight(
+            MerkleLib.getZeroRoot(),
+            MerkleLib.getMaxTreeDepth()
+        );
     }
 
     // A tree which is used in `update()` and `store()`
@@ -29,7 +33,7 @@ contract MerkleTree {
      * @param _dataBlock The data block we're storing/verifying
      * @param _path The path from the leaf to the root / the index of the leaf.
      */
-    function update(bytes memory _dataBlock, uint _path) public {
+    function update(bytes memory _dataBlock, uint256 _path) public {
         bytes32[] memory siblings = getSiblings(_path);
         store(_dataBlock, _path, siblings);
     }
@@ -39,7 +43,7 @@ contract MerkleTree {
      * @param _leaf The leaf we're storing/verifying
      *   @param _path The path from the leaf to the root / the index of the leaf.
      */
-    function updateLeaf(bytes32 _leaf, uint _path) public {
+    function updateLeaf(bytes32 _leaf, uint256 _path) public {
         bytes32[] memory siblings = getSiblings(_path);
         storeLeaf(_leaf, _path, siblings);
     }
@@ -50,7 +54,11 @@ contract MerkleTree {
      * @param _path The path from the leaf to the root / the index of the leaf.
      * @param _siblings The sibling nodes along the way.
      */
-    function verifyAndStore(bytes memory _dataBlock, uint _path, bytes32[] memory _siblings) public {
+    function verifyAndStore(
+        bytes memory _dataBlock,
+        uint256 _path,
+        bytes32[] memory _siblings
+    ) public {
         bytes32 oldRoot = tree.root;
         store(_dataBlock, _path, _siblings);
         require(tree.root == oldRoot, "Failed same root verification check!");
@@ -62,7 +70,11 @@ contract MerkleTree {
      * @param _path The path from the leaf to the root / the index of the leaf.
      * @param _siblings The sibling nodes along the way.
      */
-    function store(bytes memory _dataBlock, uint _path, bytes32[] memory _siblings) public {
+    function store(
+        bytes memory _dataBlock,
+        uint256 _path,
+        bytes32[] memory _siblings
+    ) public {
         // Compute the leaf node & store the leaf
         bytes32 leaf = keccak256(_dataBlock);
         storeLeaf(leaf, _path, _siblings);
@@ -74,13 +86,18 @@ contract MerkleTree {
      * @param _path The path from the leaf to the root / the index of the leaf.
      * @param _siblings The sibling nodes along the way.
      */
-    function storeLeaf(bytes32 _leaf, uint _path, bytes32[] memory _siblings) public {
+    function storeLeaf(bytes32 _leaf, uint256 _path, bytes32[] memory _siblings)
+        public
+    {
         // First compute the leaf node
         bytes32 computedNode = _leaf;
-        for (uint i = 0; i < _siblings.length; i++) {
+        for (uint256 i = 0; i < _siblings.length; i++) {
             bytes32 parent;
             bytes32 sibling = _siblings[i];
-            uint8 isComputedRightSibling = MerkleLib.getNthBitFromRight(_path, i);
+            uint8 isComputedRightSibling = MerkleLib.getNthBitFromRight(
+                _path,
+                i
+            );
             if (isComputedRightSibling == 0) {
                 parent = MerkleLib.getParent(computedNode, sibling);
                 // Store the node!
@@ -102,11 +119,11 @@ contract MerkleTree {
      * @param _path The path from the leaf to the root / the index of the leaf.
      * @return The sibling nodes along the way.
      */
-    function getSiblings(uint _path) public view returns (bytes32[] memory) {
+    function getSiblings(uint256 _path) public view returns (bytes32[] memory) {
         bytes32[] memory siblings = new bytes32[](tree.height);
         bytes32 computedNode = tree.root;
-        for(uint i = tree.height; i > 0; i--) {
-            uint siblingIndex = i-1;
+        for (uint256 i = tree.height; i > 0; i--) {
+            uint256 siblingIndex = i - 1;
             (bytes32 leftChild, bytes32 rightChild) = getChildren(computedNode);
             if (MerkleLib.getNthBitFromRight(_path, siblingIndex) == 0) {
                 computedNode = leftChild;
@@ -119,14 +136,12 @@ contract MerkleTree {
         // Now store everything
         return siblings;
     }
-    
+
     /**
      * @notice Append leaf will append a leaf to the end of the tree
      * @return The sibling nodes along the way.
      */
-    function appendLeaf(bytes32 leaf)public returns(bytes32){
-        
-    }
+    function appendLeaf(bytes32 leaf) public returns (bytes32) {}
 
     // }
 
@@ -137,7 +152,7 @@ contract MerkleTree {
      * @notice Get our stored tree's root
      * @return The merkle root of the tree
      */
-    function getRoot() public view returns(bytes32) {
+    function getRoot() public view returns (bytes32) {
         return tree.root;
     }
 
@@ -146,7 +161,7 @@ contract MerkleTree {
      * @param _root The merkle root of the tree
      * @param _height The height of the tree
      */
-    function setMerkleRootAndHeight(bytes32 _root, uint _height) public {
+    function setMerkleRootAndHeight(bytes32 _root, uint256 _height) public {
         tree.root = _root;
         tree.height = _height;
     }
@@ -157,7 +172,9 @@ contract MerkleTree {
      * @param _leftChild The left child of the parent in the tree
      * @param _rightChild The right child of the parent in the tree
      */
-    function storeNode(bytes32 _parent, bytes32 _leftChild, bytes32 _rightChild) public {
+    function storeNode(bytes32 _parent, bytes32 _leftChild, bytes32 _rightChild)
+        public
+    {
         tree.nodes[MerkleLib.getLeftSiblingKey(_parent)] = _leftChild;
         tree.nodes[MerkleLib.getRightSiblingKey(_parent)] = _rightChild;
     }
@@ -167,7 +184,14 @@ contract MerkleTree {
      * @param _parent The parent node
      * @return (rightChild, leftChild) -- the two children of the parent
      */
-    function getChildren(bytes32 _parent) public view returns(bytes32, bytes32) {
-        return (tree.nodes[MerkleLib.getLeftSiblingKey(_parent)], tree.nodes[MerkleLib.getRightSiblingKey(_parent)]);
+    function getChildren(bytes32 _parent)
+        public
+        view
+        returns (bytes32, bytes32)
+    {
+        return (
+            tree.nodes[MerkleLib.getLeftSiblingKey(_parent)],
+            tree.nodes[MerkleLib.getRightSiblingKey(_parent)]
+        );
     }
 }
