+++ contracts/dataTypes.sol
@@ -1,49 +1,50 @@
 pragma solidity ^0.5.0;
 pragma experimental ABIEncoderV2;
 
+
 /**
  * @title DataTypes
  * @notice TODO
  */
 contract DataTypes {
     // Batch
-    struct Batch{
+    struct Batch {
         bytes32 stateRoot;
         address committer;
         bytes32 txRoot;
-        uint stakeCommitted;
-        uint finalisesOn;
-        uint timestamp;
+        uint256 stakeCommitted;
+        uint256 finalisesOn;
+        uint256 timestamp;
     }
 
     // Account or leaf structure
-    struct Account{
-        uint balance;
-        uint tokenType;
-        uint nonce;
+    struct Account {
+        uint256 balance;
+        uint256 tokenType;
+        uint256 nonce;
     }
 
     // account leaf also contains the path to the account from the root
-    struct AccountLeaf{
-        uint path;
-        uint balance;
-        uint tokenType;
-        uint nonce;
+    struct AccountLeaf {
+        uint256 path;
+        uint256 balance;
+        uint256 tokenType;
+        uint256 nonce;
     }
 
-    struct DepositLeaf{
+    struct DepositLeaf {
         string pubkey;
     }
 
-    struct Transaction{
+    struct Transaction {
         Account from;
         Account to;
-        uint tokenType;
+        uint256 tokenType;
         uint32 amount;
         bytes signature;
     }
 
-    struct MerkleProof{
+    struct MerkleProof {
         AccountLeaf account;
         bytes32[] siblings;
     }
